refactor(client): update clipboard component
Remove the `react-clipboard.js` library.
