+++ client/package.json
@@ -38,7 +38,6 @@
     "react-autosuggest": "^10.1.0",
     "react-avatar-editor": "^13.0.0",
     "react-bootstrap-icons": "^1.10.3",
-    "react-clipboard.js": "^2.0.16",
     "react-collapse": "^5.0.0",
     "react-cookie-consent": "^8.0.1",
     "react-dom": "^18.2.0",
